changed some types

+++ src/main/java/edu/brown/cs/term_project/TextSimilarity/IText.java
@@ -9,5 +9,5 @@ public interface IText<W extends IWord> {
    * @param textType and integer telling which type of text (ie Entity  vs Vocab)
    * @return as hashmap that maps from word to frequency in text.
    */
-  Map<W, Double> getFreq(Integer textType);
+  Map<W, Integer> getFreq(Integer textType);
 }
